+++ app/controllers/panel/services/controller.go
@@ -1,0 +1,22 @@
+package services
+
+import (
+	"context"
+	"net/http"
+
+	"golazy.dev/lazy/app/controllers/panel"
+)
+
+type Controller struct {
+	panel.Base
+}
+
+func New(ctx context.Context) (*Controller, error) {
+	base, err := panel.NewBase(ctx)
+	return &Controller{Base: base}, err
+}
+
+func (c *Controller) Index(_ http.ResponseWriter, _ *http.Request) error {
+	c.SetState()
+	return nil
+}